youtube_travel_analyzer/requirements.txt
pandas
google-api-python-client
konlpy
scikit-learn
python-dotenv